refactor(search): scripts and lib into shared dir
