+++ snippets/src/libs/build-search-index.ts
@@ -1,5 +1,5 @@
 import { executeBuildSearchIndex } from "./search-libs";
-import { pinoLogBuilder } from "./utils/pino-log";
-import { logTime } from "./utils/utils";
+import { pinoLogBuilder } from "../utils/pino-log";
+import { logTime } from "../utils/utils";
 
 logTime('build-search-index', executeBuildSearchIndex, pinoLogBuilder('main', 'info'));
