+++ snippets/src/libs/type.ts
