+++ snippets/src/libs/uuid.ts
