+++ snippets/src/scripts/add-id.ts
@@ -1,9 +1,9 @@
-import { MarkdownFileProcessor, processMarkdownDirectories } from "./libs/markdown-files";
+import { MarkdownFileProcessor, processMarkdownDirectories } from "../libs/markdown-files";
 import { config } from "./_config";
 import { generateIdMapper } from "./generate-id-mapper";
 import fs from 'fs-extra';
 import * as core from '@actions/core';
-import { pinoLogBuilder } from "./utils/pino-log";
+import { pinoLogBuilder } from "../utils/pino-log";
 
 const { sourceDirectories, ignoreMarkdownFiles } = config.blogIdModule;
