+++ snippets/src/utils/utils.ts
